assn3: method header update

+++ libsponge/tcp_sender.cc
@@ -22,7 +22,7 @@ TCPSender::TCPSender(const size_t capacity, const uint16_t retx_timeout, const s
     , _initial_retransmission_timeout{retx_timeout}
     , _stream(capacity) {}
 
-uint64_t TCPSender::bytes_in_flight() const { return {}; }
+size_t TCPSender::bytes_in_flight() const { return {}; }
 
 void TCPSender::fill_window() {}
